api calls & more options dialog
