+++ src/frontend/staff-portal/src/app/components/jj-count/jj-count.component.scss
@@ -16,20 +16,20 @@ h3 {
 }
 
 .section-grid-cell {
-  border:none; 
+  border:none;
   padding-left: 5px;
   padding-bottom: 15px;
 }
 
 .section-jj-grid-cell {
-  border-right-width: 2px; 
+  border-right-width: 2px;
   border-right-style: solid;
   border-right-color: #e4f2ff;
   padding-left: 5px;
   padding-bottom: 15px;
 }
 .section-grid-header {
-  font-size:small; 
+  font-size:small;
   color: #707070;
   font-weight: bold;
   margin: 0 !important;
@@ -37,14 +37,14 @@ h3 {
 }
 
 .section-grid-header2 {
-  font-size:small; 
+  font-size:small;
   color: #707070;
   margin: 0 !important;
   padding: 0 !important;
 }
 
 .section-grid-text {
-  font-size:small; 
+  font-size:small;
   font-weight: bold;
   color: #1d4a77 !important;
   margin: 0 !important;
@@ -59,7 +59,7 @@ h3 {
 }
 
 .section-jj-grid-text {
-  font-size:small; 
+  font-size:small;
   color: #1d4a77 !important;
   margin: 0 !important;
   padding: 0 !important;
@@ -78,7 +78,7 @@ p {
 }
 
 .button {
-  padding: 20px;
+  padding: 10px;
   background-color: #c2c2c2;
   color: #fff;
 }
@@ -91,10 +91,9 @@ p {
   }
 
 .mat-flat-button {
-  width: 30% !important;
   float: right;
   margin-top: 1em;
-  margin-right: 20px;
+  margin-right: 10px;
 
   &.large {
     height: 46px;
@@ -126,7 +125,7 @@ p {
     }
     // :host
   }
-  
+
   mat-expansion-panel {
     box-shadow: none !important;
     padding-bottom: 1em;
