+++ src/frontend/staff-portal/src/app/components/jj-dispute/jj-dispute.component.html
@@ -138,9 +138,6 @@
       <button class="large" mat-flat-button type="button" color="primary" (click)="onSubmit()" [disabled]="lastUpdatedJJDispute.status != 'NEW' && lastUpdatedJJDispute.status != 'IN_PROGRESS'">
         Submit
       </button>
-      <button class="large" mat-flat-button type="button" color="primary" (click)="onMoreOptions()"  [disabled]="lastUpdatedJJDispute.status != 'NEW' && lastUpdatedJJDispute.status != 'IN_PROGRESS'">
-        More Options
-      </button>
       <button class="large" mat-flat-button type="button" color="primary" (click)="onSave()"  [disabled]="lastUpdatedJJDispute.status != 'NEW' && lastUpdatedJJDispute.status != 'IN_PROGRESS'">
         Save & Finish Later
       </button>
